Add DialogProgressBG() messages on status of power commands
As it takes around 1 second for the CEC status command to complete, and
then a little longer for the actual turn off TV command to execute, the
power button feels very sluggish, and users can assume that their
keypress may not have been registered. A dialogue message to communicate
that something is happening helps remedy this.